ADD: Timed slides option. Add 'data-time' with the milliseconds the slide (including transition animation) should stay on screen for until moving onto the next slide.
ADD: Time till next slide bar which should (though needs tweaking) move across the screen to 100% of the screen width when the slide is about to move onto the next. Timebar doesn't show up on slides without a data-time option.

REM: Removed the hash update listener for the time being. It was resulting in slides loading twice each time which was throwing off the timer function.

Known Issues:
If you skip a slide with the keyboard the timer bar will rush across the screen too quickly. Need to resolve.

Need to fix the usage of the hash update listener so the timer can work with the listener enabled.