Implemented a simple API for intercepting navigation.
This change allows developers to provide custom code that runs when navigating from one step to the next. The following is an example of how a navigation handler might be setup within the top of the html page of the presentation.

    <script>
        window.Impress = {
            beforeNavigation: function(before, next) {
                if (before != null && before.id == 'slide-1' && next.id == 'slide-2') {
                    alert('yes!');
                }
            },
            afterNavigation: function(before, current) {}
        };
    </script>

Each method provides the before element and the element we're navigating to. This allows the developer to determine whether they're moving forward or backward. But also lets them transition out the old element and transition the new one if need be.