Skippable positions
It's nice to be able to remain consistent with the markup for background stuff but maybe you don't want it in a slide.  Now you can add .skip to an element and it will be positioned but not scrolled to.  Maybe a little hacky but it gets the job done right!