Update index.html
added demos on how to use the inner step. Don't know if you like this method but I thought it would offer the most functionality. Using an array of json objects in the data-steps prop you can specify the css you want for each step. This isn't incredibly interesting demo. But I could make one and that is the happy part.