Adding concept of a "Key" step
I added the concept of a "Key" step that can be easily referenced from
anywhere in the presentation. Often during presentations you have a
slide that provides an overview the topic or contains relevant
information that you want to refer back to often. This allows you to set
that "Key" step by adding a "keyStep" class to the step's div element.
Then you can press the "k" key from anywhere in the presentation to go
to/from the key step to wherever you currently are in the steps (works
for going forward or backwards).