init() reset method
Initially forked the impress.js repo to create dependency for our
project. We could not init() consecutively without trouble; impress
would not setup the presentation correctly on the second init(). Likely
because roots is not flushed on init() and so remains the second time
with stale data. This change seems to fix it.