Feature: Substeps
We needed the "substep"-feature mentioned in the
https://github.com/bartaz/impress.js/issues/81 issue.

We implemented 4 API functions for assigning functions to the steps.
Every mentioned API function takes two parameters: The id of the step
and a function that should be executed on a given time.

There can be one function per step for entering.
There can be one function per step for leaving.
There can be many functions per step that act as "substeps". The user
can then modify everything he wants with simple js. These functions are
called trough next(). The first defined function is executed first.
There can be one function per step that allows resetting the changes
done by the "substep-functions".

Our changes are all backwards-compatible, so it will not break anything
from anyone. We have tested it with a variety of browsers.

We did this with 13 line html (demo), 10 line css (demo) and 46 lines js
(framework). In total less than 70 lines of code added.

If you have something you want to speak about or you are not happy with,
feel free to contact us (info@patricebeeutler.ch).