Add support for relative positioning of steps.
Adds support for following new attributes for steps:
  data-rel-x
  data-rel-y
  data-rel-z

These can be used to specify position instead of the
absolute position arguments data-x/y/z.

Incompatible change: Previously the default value
of data-x/y/z was 0. Now the default is to inherit
their value from the previous step. The default for the
initial step is still 0.

The value of data-rel-x/y/z is added to data-x/y/z. If not
specified, also the rel-* values are inherited from the
previous step. This makes it easy to create a boring presentation
where each slide shifts for example 1000px down from the previous.

Note: In https://github.com/bartaz/impress.js/issues/435 the
impress.js community is discussing about moving to a more plugin
based architecture. If that happens, this patch could be added
as some kind of "positioning plugin", who's job is to transform
higher order attributes into the absolute x/y/z values that exist
in the core code.

If such a plugin architecture is introduced in the future, it's
simply to extract out this functionality at that time (and could
be a useful implementation of a first plugin for positioning, as
this is a very trivial yet useful addition).