Add navigation-ui plugin.
navigation-ui plugin provides "back" and "forward" controls,
as well as a select drop down list to jump to any step. It is
added to act as an example of a UI plugin, meaning that it
exposes visible html elements.

(This plugin depends on the toolbar plugin, which is added 3 commits
from now.)